Organised the fitting software folder and fixed some redundancies
