+++ fitbenchmarking/fitting/mantid/__init__.py
@@ -1,0 +1,1 @@
+
