+++ fitbenchmarking/fitting/mantid/externals.py
@@ -1,5 +1,5 @@
 """
-Parse the problem file depending on the type of problem.
+Fitting and utility functions for the mantid fitting software.
 """
 # Copyright &copy; 2016 ISIS Rutherford Appleton Laboratory, NScD
 # Oak Ridge National Laboratory & European Spallation Source
@@ -24,35 +24,33 @@ Parse the problem file depending on the type of problem.
 
 from __future__ import (absolute_import, division, print_function)
 
-from parsing import parse_nist, parse_neutron
+import mantid.simpleapi as msapi
 from utils.logging_setup import logger
 
 
-def parse_problem_file(group_name, prob_file):
+def gen_func_obj(function_name):
     """
-    Helper function that does the parsing of a specified problem file.
-    This method needs group_name to inform how the prob_file should be
-    passed.
+    Generates a mantid function object.
 
-    @param group_name :: name of the group of problems
-    @param prob_file :: path to the problem file
+    @param function_name :: the name of the function to be generated
 
-    @returns :: problem object with fitting information
+    @returns :: mantid function object that can be called in python
     """
+    exec "function_object = msapi." + function_name + "()"
+    return function_object
 
-    if group_name == 'nist':
-        prob = parse_nist.load_file(prob_file)
-        prob.type = 'nist'
-    elif group_name == 'neutron':
-        prob = parse_neutron.load_file(prob_file)
-        prob.type = 'neutron'
-    # elif ...
-    #    prob = call_parse_function_here
-    #    prob.type = ...
-    else:
-        raise NameError("Could not find group name! Please check if it was"
-                        "given correctly...")
 
-    logger.info("* Testing fitting of problem {0}".format(prob.name))
+def set_ties(function_object, ties):
+    """
+    Sets the ties for a function/composite function object.
 
-    return prob
+    @param function_object :: mantid function object
+    @param ties :: array of strings containing the ties
+
+    @returns :: mantid function object with ties
+    """
+    for idx, ties_per_func in enumerate(ties):
+        for tie in ties_per_func:
+            exec "function_object.tie({'f" + str(idx) + "." + tie + "})"
+
+    return function_object
