+++ fitbenchmarking/fitting/mantid/tests/test_mantid.py
