+++ fitbenchmarking/fitting/prerequisites.py
@@ -25,8 +25,7 @@ specific.
 
 from __future__ import (absolute_import, division, print_function)
 
-from fitting import mantid
-from fitting import scipyfit
+import fitting
 import numpy as np
 
 
@@ -54,16 +53,35 @@ def prepare_software_prerequisites(software, problem, use_errors):
 
 
 def prepare_mantid(problem, use_errors):
+    """
+    Does the things specified in the above description for benchmarking
+    the mantid software.
 
-    wks_mtd, cost_function = mantid.wks_cost_function(problem, use_errors)
-    function_definitions = mantid.function_definitions(problem)
+    @returns :: mantid workspace data structure
+                mantid cost funtion
+                mantid function definitions
+    """
+    wks_mtd, cost_function = \
+    fitting.mantid.prepare_data.wks_cost_function(problem, use_errors)
+    function_definitions = \
+    fitting.mantid.func_def.function_definitions(problem)
+
     return wks_mtd, cost_function, function_definitions
 
 
 def prepare_scipy(problem, use_errors):
+    """
+    Does the things specified in the above description for benchmarking
+    the scipy software.
 
-    data, cost_function = scipyfit.prepare_data(problem, use_errors)
-    function_definitions = scipyfit.function_definitions(problem)
+    @returns :: scipy workspace data structure
+                scipy cost funtion
+                scipy function definitions
+    """
+    data, cost_function = \
+    fitting.scipy.prepare_data.prepare_data(problem, use_errors)
+    function_definitions = fitting.scipy.func_def.function_definitions(problem)
+    # For problems that have no specified boundries, set -inf and +inf
     if problem.start_x == None and problem.end_x == None:
         problem.start_x = - np.inf
         problem.end_x = np.inf
