+++ fitbenchmarking/fitting/scipy/tests/test_scipyfit.py
